moving error package to root exception
